Move collection page state to common
