+++ packages/web/src/common/store/pages/collection/lineup/reducer.js
@@ -1,6 +1,6 @@
 import { RESET_SUCCEEDED, stripPrefix } from 'common/store/lineup/actions'
 import { initialLineupState } from 'common/store/lineup/reducer'
-import { PREFIX } from 'pages/collection-page/store/lineups/tracks/actions'
+import { PREFIX } from 'common/store/pages/collection/lineup/actions'
 
 const initialState = {
   ...initialLineupState,
