+++ packages/web/src/pages/smart-collection/store/sagas.ts
@@ -6,7 +6,7 @@ import { Track, UserTrack } from 'common/models/Track'
 import { getAccountStatus, getUserId } from 'common/store/account/selectors'
 import { processAndCacheTracks } from 'common/store/cache/tracks/utils'
 import { fetchUsers as retrieveUsers } from 'common/store/cache/users/sagas'
-import { setSmartCollection } from 'pages/collection-page/store/actions'
+import { setSmartCollection } from 'common/store/pages/collection/actions'
 import Explore from 'services/audius-backend/Explore'
 import { waitForBackendSetup } from 'store/backend/sagas'
 import { getLuckyTracks } from 'store/recommendation/sagas'
